finished (i think) thornbella
dust trail
swoosh!
etc.

+++ Projectiles/MortJavelin/MortJavelin.cs
@@ -11,8 +11,10 @@ namespace MythosOfMoonlight.Projectiles.MortJavelin
     {
         public override void SetStaticDefaults()
         {
-            DisplayName.SetDefault("Thornbella");
+            DisplayName.SetDefault("Mortiflora Javelin");
             Main.projFrames[projectile.type] = 1;
+            ProjectileID.Sets.TrailingMode[projectile.type] = 2;
+            ProjectileID.Sets.TrailCacheLength[projectile.type] = 40;
         }
 
         public override void SetDefaults()
@@ -27,9 +29,11 @@ namespace MythosOfMoonlight.Projectiles.MortJavelin
 
         private const int MAX_TICKS = 45;
 
+        private float velocity = 12f;
         private void NormalAI()
         {
             TargetWhoAmI++;
+            timeSinceLastBounced++;
 
             // For a little while, the javelin will travel with the same speed, but after this, the javelin drops velocity very quickly.
             if (TargetWhoAmI >= MAX_TICKS)
@@ -52,34 +56,55 @@ namespace MythosOfMoonlight.Projectiles.MortJavelin
             foreach (NPC npc in Main.npc)
             {
                 float potentialDistance = Vector2.Distance(projectile.position, npc.Center);
-                if ((npc.friendly || potentialDistance > 81 * BLOCK_LENGTH) || potentialDistance >= closestDistance)
+                if ((npc.friendly || !npc.active || potentialDistance > 48 * BLOCK_LENGTH) 
+                    || potentialDistance >= closestDistance)
                     continue;
-                closestTarget = npc;
+                bool collidesWith = Collision.CanHitLine(projectile.position, projectile.width, projectile.height, npc.position, npc.width, npc.height);
+                if (collidesWith)
+                {
+                    closestTarget = npc;
+                    closestDistance = potentialDistance;
+                }
             }
+            Main.NewText(closestDistance.ToString());
+            if (closestDistance <= 16 * BLOCK_LENGTH)
+                projectile.tileCollide = false;
             return closestTarget;
         }
 
         private void Bounce()
         {
             target = GetTarget();
-            bool hitTarget = false;
+
             if (target != null)
             {
-                if (Collision.CanHitLine(projectile.position, projectile.width, projectile.height, target.position, target.width, target.height))
+                // if ()
+                // {
+                Main.PlaySound(SoundID.Item71);
+                var direction = -(projectile.Center - target.Center);
+                direction.Normalize();
+                projectile.velocity = direction * (velocity * 2);
+
+                //doost
+                Dust dust;
+                Vector2 position = projectile.Center;
+                for (int i = 0; i < 17; i++)
                 {
-                    var direction = -(projectile.position - target.position);
-                    direction.Normalize();
-                    projectile.velocity = direction * Magnitude(projectile.velocity.RotatedBy(15));
+                    dust = Main.dust[Terraria.Dust.NewDust(position, 20, 20, 1, 0f, 0f, 0, new Color(255, 0, 0), 1f)];
+                    dust.fadeIn = 0f;
                 }
 
-                else
-                {
-                    hitTarget = true;
-                }
+                // }
+
+                // else
+                // {
+                //     hitTarget = true;
+                // }
             }
 
-            if (hitTarget)
+            else
             {
+                Main.PlaySound(SoundID.NPCHit2);
                 projectile.velocity = new Vector2(projectile.velocity.X, -projectile.velocity.Y);
             }
 
@@ -87,7 +112,6 @@ namespace MythosOfMoonlight.Projectiles.MortJavelin
             // {
             //    projectile.Kill();
             // }
-            TargetWhoAmI = 0;
         }
 
         private float Magnitude(Vector2 mag)
@@ -121,7 +145,10 @@ namespace MythosOfMoonlight.Projectiles.MortJavelin
                     Bounce();
                     break;
                 default:
-                    projectile.Kill();
+                    if (timeSinceLastBounced > 2)
+                        projectile.Kill();
+                    else
+                        Bounce();
                     break;
             }
             return false;
@@ -139,9 +166,48 @@ namespace MythosOfMoonlight.Projectiles.MortJavelin
             set => projectile.ai[1] = value;
         }
 
+        int timeSinceLastBounced = 0;
+
+        public override bool PreDraw(SpriteBatch spriteBatch, Color lightColor)
+        {
+            //3hi31mg
+            var clr = new Color(255, 255, 255, 255); // full white
+            var drawPos = projectile.Center - Main.screenPosition;
+            var off = new Vector2(projectile.width / 2f, projectile.height / 2f);
+            var origTexture = Main.projectileTexture[projectile.type];
+            var texture = mod.GetTexture("Projectiles/MortJavelin/MortJavelinGlow");
+            int frameHeight = texture.Height / Main.projFrames[projectile.type];
+            var frame = new Rectangle(0, frameHeight * projectile.frame, texture.Width, frameHeight - 2);
+            var orig = frame.Size() / 2f;
+
+            Main.spriteBatch.Draw(origTexture, drawPos, frame, clr, projectile.rotation, orig, projectile.scale, SpriteEffects.None, 0f);
+            Main.spriteBatch.Draw(texture, drawPos, frame, clr, projectile.rotation, orig, projectile.scale, SpriteEffects.None, 0f);
+
+            var fadeMult = 1f / ProjectileID.Sets.TrailCacheLength[projectile.type];
+            off -= new Vector2(0, 9).RotatedBy(projectile.rotation);
+            var dust = Terraria.Dust.NewDustPerfect(projectile.position + off + (projectile.velocity * 2), 235, new Vector2(), 0, new Color(255, 255, 255), 1f);
+
+            if (target != null)
+            {
+                var red = new Color(255, 0, 0, 155);
+                for (int i = 1; i < ProjectileID.Sets.TrailCacheLength[projectile.type]; i++)
+                {
+                    Main.spriteBatch.Draw(texture, projectile.oldPos[i] - Main.screenPosition, frame, red * (1f - fadeMult * i), projectile.oldRot[i], orig, projectile.scale, SpriteEffects.None, 0f);
+                }
+            }
+            return false;
+        }
+
         private NPC target;
         public override void AI()
         {
+            Dust dust;
+            // You need to set position depending on what you are doing. You may need to subtract width/2 and height/2 as well to center the spawn rectangle.
+            Vector2 position = projectile.Center + new Vector2(-2, 32).RotatedBy(projectile.rotation);
+            dust = Terraria.Dust.NewDustPerfect(position, 1, new Vector2(0f, 0f), 0, new Color(177, 255, 0), 1f);
+            dust.noGravity = true;
+            dust.fadeIn = 0.9f;
+
             NormalAI();
         }
     }
